+++ engine/flink/management/sample/src/main/scala/pl/touk/nussknacker/engine/management/sample/signal/RemoveLockProcessSignalFactory.scala
@@ -2,7 +2,6 @@ package pl.touk.nussknacker.engine.management.sample.signal
 
 import java.lang
 import java.nio.charset.StandardCharsets
-
 import com.typesafe.scalalogging.LazyLogging
 import io.circe.{Encoder, Json}
 import io.circe.generic.JsonCodec
@@ -17,7 +16,7 @@ import pl.touk.nussknacker.engine.api.{MethodToInvoke, ParamName, _}
 import pl.touk.nussknacker.engine.flink.api.process.{FlinkCustomNodeContext, FlinkCustomStreamTransformation}
 import pl.touk.nussknacker.engine.flink.api.signal.FlinkProcessSignalSender
 import pl.touk.nussknacker.engine.flink.util.signal.KafkaSignalStreamConnector
-import pl.touk.nussknacker.engine.kafka.{KafkaConfig, KafkaUtils}
+import pl.touk.nussknacker.engine.kafka.{DefaultProducerCreator, KafkaConfig, KafkaUtils}
 
 
 class RemoveLockProcessSignalFactory(val kafkaConfig: KafkaConfig, val signalsTopic: String)
@@ -28,7 +27,7 @@ class RemoveLockProcessSignalFactory(val kafkaConfig: KafkaConfig, val signalsTo
   @MethodToInvoke
   def sendSignal(@ParamName("lockId") lockId: String)(processId: String): Unit = {
     val signal = SampleProcessSignal(processId, System.currentTimeMillis(), RemoveLock(lockId))
-    KafkaUtils.sendToKafkaWithTempProducer(signalsTopic, Array.empty, Encoder[SampleProcessSignal].apply(signal).noSpaces.getBytes(StandardCharsets.UTF_8))(kafkaConfig)
+    KafkaUtils.sendToKafkaWithTempProducer(signalsTopic, Array.empty, Encoder[SampleProcessSignal].apply(signal).noSpaces.getBytes(StandardCharsets.UTF_8))(DefaultProducerCreator(kafkaConfig))
   }
 
 }